backend/config/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('', include('web.urls')),
    path('admin/', admin.site.urls),
    path('api/v1/auth/token/', TokenObtainPairView.as_view(), name='token_obtenir'),
    path('api/v1/auth/token/rafraichir/', TokenRefreshView.as_view(), name='token_rafraichir'),
    path('api/v1/utilisateurs/', include('users.urls')),
    path('api/v1/projets/', include('projects.urls')),
    path('api/v1/taches/', include('tasks.urls')),
    path('api/v1/statistiques/', include('statistiques.urls')),
    path('accounts/', include('allauth.urls')),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
